+++ CHANGES.md
@@ -1,4 +1,4 @@
-# `next` branch
+# 0.25.0 / 2020-05-20
 
 ## Breaking Changes